interface OtherInfoDataBase
creada interface OtherInfoDataBase
y actualizada OtherInfoDataBaseImpl
